+++ src/Symfony/Component/Routing/Loader/PhpFileLoader.php
@@ -22,6 +22,8 @@ use Symfony\Component\Routing\RouteCollection;
  * The file must return a RouteCollection instance.
  *
  * @author Fabien Potencier <fabien@symfony.com>
+ * @author Nicolas grekas <p@tchwork.com>
+ * @author Jules Pietri <jules@heahprod.com>
  */
 class PhpFileLoader extends FileLoader
 {
@@ -47,8 +49,7 @@ class PhpFileLoader extends FileLoader
         $result = $load($path);
 
         if (\is_object($result) && \is_callable($result)) {
-            $collection = new RouteCollection();
-            $result(new RoutingConfigurator($collection, $this, $path, $file));
+            $collection = $this->callConfigurator($result, $path, $file);
         } else {
             $collection = $result;
         }
@@ -64,6 +65,15 @@ class PhpFileLoader extends FileLoader
     public function supports($resource, string $type = null)
     {
         return \is_string($resource) && 'php' === pathinfo($resource, PATHINFO_EXTENSION) && (!$type || 'php' === $type);
+    }
+
+    protected function callConfigurator(callable $result, string $path, string $file): RouteCollection
+    {
+        $collection = new RouteCollection();
+
+        $result(new RoutingConfigurator($collection, $this, $path, $file));
+
+        return $collection;
     }
 }